proxy loads config from json and creates backends

+++ hacking/stop.sh
@@ -1,7 +1,8 @@
 #!/bin/bash -exu
 
-SYMFONY_BIN=./symfony-cli
-PROXY_JSON=~/.symfony5/proxy.json
+SYMFONY_BIN="./symfony-cli"
+PROXY_JSON="$HOME/.symfony5/proxy.json"
+PROXY_JSON_DELETE=false
 
 $SYMFONY_BIN proxy:stop
 $SYMFONY_BIN proxy:status
@@ -12,7 +13,7 @@ $SYMFONY_BIN server:stop
 if [[ -f "$PROXY_JSON" ]]
 then
     cat $PROXY_JSON
-    rm $PROXY_JSON
+    $PROXY_JSON_DELETE && rm $PROXY_JSON
 fi
 
 echo
